+++ publish-schema/src/pub_sub_error_message.py
@@ -1,8 +1,9 @@
 import base64
 import json
+import logging
+
+from config import CONFIG
 from google.cloud import pubsub_v1
-from config import config
-from logging_config import logging
 
 logger = logging.getLogger(__name__)
 
@@ -13,7 +14,7 @@ class PubSubErrorMessage:
         self.error_message = error_message
         self.schema_file = schema_file
 
-        self.send_message(config.PUBLISH_SCHEMA_ERROR_TOPIC_ID)
+        self.send_message(CONFIG.PUBLISH_SCHEMA_ERROR_TOPIC_ID)
 
     def to_json(self) -> str:
         """
@@ -35,7 +36,7 @@ class PubSubErrorMessage:
             dict: The message sent to the Pub/Sub topic.
         """
         publisher = pubsub_v1.PublisherClient()
-        topic_path = publisher.topic_path(config.PROJECT_ID, topic)
+        topic_path = publisher.topic_path(CONFIG.PROJECT_ID, topic)
         message = self.build_pubsub_message()
         future = publisher.publish(topic_path, message.encode("utf-8"))
         future.result()
